SupplierList + Update to Supplier
Added SupplierList which includes an iterator that will go through each Supplier in the list.
To give the iterator a use also added toString Methods to both SupplierList and Supplier.
Supplier's viewProductList now returns a String to allow a list of products to be printed when supplier.toString is called